EchoCommon elastic client#new has equal arguments as elastic gem.
This way it is easier to pass in other options than only configuration
for the host. We can also fill in request timeouts etc as if it is the
"native" elastic search gem's client.

We also delete arguments our client cares about so they don't get in to
the elastic's client new argument list.