Added class representing one-to-many-relation.
A bit unsure if this class deserves it’s right to exist, as it does not
do much. We do have a couple of places in Echo where we have objects
representing relation, so to have the same basic public API for these
abstraction I extracted this code in to this class.