#authenticated? only rescues JWT ExpiredSignature
Instead of all decode errors. My first thought was that we could rescue
all decode errors, but maybe it is better to simply select this one and
see how that goes and if we get other types of errors as well. The
advantage with this is that we don’t mask any real errors in the
returned false without logging it. Now at least we’ll know.