Trying to figure one-to-one types for Hanami::Entities.
The call to Types::Schema::CoercibleType.new seems to be a bit
internal, as it has not been exposed like Types::Collection has been.
It works, but only due to the fact that I removed the to_hash method to
stop symbolize of attrs hash which Hanami does from hashify the correct
type.