Add support for search with indexes named 'foo,bar'.
I added a tiny bit of checking on the index name, to be clear on what we
expect/support, and what we don't.

Like, we don't support '-foo', to express not foo index, as we need to
do something begtter for stuff like 'foos*,-foos3'. In this case we want
get out: 'testing_foos*,-testing_foos3' (in other words move the - in
front of the prefix.