Duration supports "P0DT4M4S", instead of "PT4M4S".
Now, I guess to fully support this we need to rework our code a bit, as
we still don't support for instance "P1DT2H".

And we would return "PT26H" for the duration aboive, which I don't think
is the correct/allowed thing to do.

On the other side. What kind of a recording lasts for more than 24
hours?