feat: added an elasticsearch restore wait-on-completion helper.
We need to query the recovery api for the cluster in order to figure out
if we have completed and restored every index.

We found no way to figure this out via elasticsearch's task api, as it
only seems to be listed there if the restore request itself has
wait_for_completion=true. We need to wait longer than the gateway of
cloud.elastic.co allows in order to restore production indices, so this
code allows us to wait by polling.