fix: include id in returned elastic doc, for search, get and mget
Soo, previously, we actually indexed the id of the doc in an
id field in the doc's source. We don't do that any more, but the
callee still expects this behaviour to find the id in the doc's source.

So this change makes restores the behaviour we had, even though we don't
actually store the id in the doc's source any more.