feat(elastic): suppress shards failures when searching
Our reason for requiring this is that we have repositories in Echo,
which works towards recordings index to find recordings with given
producer owner ids.

The thing which complicates it is that this recording repository
actually works towards a @discography alias which spans over recordings
and discography_drafts index.

The draft index does not have the same fields as the recording in
regards to for instance ownerships.

So when we introduced the IndexShardsError error it made this code in
Echo fail.

The IndexShardsError was introduced back in October 2019, but didn't
make its way in to Echo before now :-(