Add "external_foreign_key" to recordings (#111)
Currently we rely on local ids in echo to store the connection between an echo
recording and the draft it is based on. The problem is that we "retire" entries
when new data about a release/recording comes in, and we would like to "clean"
that old data without breaking the connection to echo.

The plan is to actually replace the ID in ES with the external foreign key, but
since release already had this field I figured I may as well add it anyway, we
can always remove it later. I also "fixed" the types we use to be `keyword`.

Connected to https://github.com/gramo-org/echo/issues/5434