Use a set in the `Entity::Relation::Many` class for much better performance (#124)
When we have large collections, using `include?` requires checking the whole
array every time a new item is added, which kill performance. Using a `Set`
means it takes a fraction of the time.

Also, `eq?` is not a method on arrays, so I have removed it.

Connected to https://github.com/gramo-org/echo/issues/6287